refactor: rename domains
