+++ applications/deposit.ts
@@ -2,12 +2,13 @@ import { DEPOSIT_TYPE_ONE_TIME } from "../lib/constants";
 
 import Queue from "../lib/dataStructures/queue";
 import { buildBalanceObject } from "../domains/balance";
+import type { DepositPlan } from "../domains/depositPlan";
+import type { Deposit } from "../domains/deposit";
 import {
   checkDepositPlansConstraints,
-  DepositPlan,
   sortDepositPlans,
-} from "../domains/depositPlans";
-import { checkDepositsConstraints, Deposit, getSum } from "../domains/deposits";
+} from "../domains/depositPlan";
+import { checkDepositsConstraints, getSum } from "../domains/deposit";
 
 /**
  * @description
